omni_binding_plugin/requirements.txt
# USD Binding Protocol Extension Requirements

# Core dependencies (installed by Omniverse)
# omni.kit.usd
# omni.usd
# omni.kit.property.usd
# omni.kit.window.property
# omni.kit.commands
# omni.timeline
# omni.ui

# Optional protocol-specific dependencies
# Install these based on which protocols you need:

# For MQTT support
aiomqtt>=2.0.0

# For HTTP/REST support
aiohttp>=3.8.0

# For SQL support
aioodbc>=0.4.0

# For gRPC support
grpcio>=1.50.0

# For WebSocket support
websockets>=11.0.0

# Note: File protocol has no additional dependencies (uses built-in file operations)
